move yml logic to common
