+++ internal/common/utilities.go
@@ -3,7 +3,9 @@ package common
 import (
 	"bytes"
 	"strings"
+	"testing"
 
+	"github.com/stretchr/testify/assert"
 	yaml "gopkg.in/yaml.v3"
 )
 
@@ -27,4 +29,30 @@ func TrustedUnmarshalYAML(d string) map[string]interface{} {
 		panic(err)
 	}
 	return parsedYaml
+}
+
+func YmlUnmarshall(in string, out interface{}) error {
+	err := yaml.Unmarshal([]byte(in), out)
+	return err
+}
+
+func YmlUnmarshalTestHelper(input string, out any, t *testing.T) {
+	t.Helper()
+	err := YmlUnmarshall(input, out)
+	assert.NoError(t, err)
+}
+
+func YmlUnmarshalMap(input string, t *testing.T) map[string]interface{} {
+	t.Helper()
+	var data map[string]interface{}
+	err := yaml.Unmarshal([]byte(input), &data)
+	assert.NoError(t, err)
+	return data
+}
+
+func YmlMarshal(input interface{}, t *testing.T) string {
+	t.Helper()
+	data, err := yaml.Marshal(&input)
+	assert.NoError(t, err)
+	return string(data)
 }
